LCH:6076: Removing trait.
