+++ src/Command/ContentHubAudit.php
@@ -17,12 +17,22 @@ use Symfony\Component\Console\Output\OutputInterface;
 class ContentHubAudit extends Command implements PlatformBootStrapCommandInterface {
 
   use CommandExecutionTrait;
-  use ContentHubModuleFunctionExplorerTrait;
 
   /**
    * {@inheritdoc}
    */
   protected static $defaultName = 'ach:audit:full';
+
+  public const V1_MODULE_HOOKS = [
+    'acquia_contenthub_drupal_to_cdf_alter',
+    'acquia_contenthub_cdf_from_drupal_alter',
+    'acquia_contenthub_cdf_from_hub_alter',
+    'acquia_contenthub_drupal_from_cdf_alter',
+    'acquia_contenthub_exclude_fields_alter',
+    'acquia_contenthub_field_type_mapping_alter',
+    'acquia_contenthub_cdf_alter',
+    'acquia_contenthub_is_eligible_entity',
+  ];
 
   /**
    * {@inheritdoc}
@@ -126,7 +136,7 @@ class ContentHubAudit extends Command implements PlatformBootStrapCommandInterfa
       foreach ($regex as $module_file) {
         $functions = $this->getModuleFunctions($module_file[0]);
         $file_info = pathinfo($module_file[0]);
-        foreach ($this->hooks as $hook) {
+        foreach (self::V1_MODULE_HOOKS as $hook) {
           if (array_search("{$file_info['filename']}_$hook", $functions) !== FALSE) {
             $ok = FALSE;
             $output->writeln(sprintf("The %s module implements ContentHub 1.x hook %s and must be converted before upgrading to 2.x.", $file_info['filename'], $hook));
@@ -138,6 +148,72 @@ class ContentHubAudit extends Command implements PlatformBootStrapCommandInterfa
       // Let's style this so it looks awesome.
       $output->writeln("<info>No deprecated ContentHub 1.x API implementations detected.</info>");
     }
+  }
+
+  /**
+   * Tokenizes module functions.
+   *
+   * @todo considering genericizing this in another class.
+   *
+   * @param string $file
+   *   The file to tokenize.
+   *
+   * @return array
+   *   Array.
+   */
+  protected function getModuleFunctions(string $file): array {
+    $source = file_get_contents($file);
+    $tokens = token_get_all($source);
+
+    $functions = [];
+    $nextStringIsFunc = FALSE;
+    $inClass = FALSE;
+    $bracesCount = 0;
+
+    foreach ($tokens as $token) {
+      switch ($token[0]) {
+        case T_CLASS:
+          $inClass = TRUE;
+          break;
+
+        case T_FUNCTION:
+          if (!$inClass) {
+            $nextStringIsFunc = TRUE;
+          }
+          break;
+
+        case T_STRING:
+          if ($nextStringIsFunc) {
+            $nextStringIsFunc = FALSE;
+            $functions[] = $token[1];
+          }
+          break;
+
+        // Anonymous functions.
+        case '(':
+        case ';':
+          $nextStringIsFunc = FALSE;
+          break;
+
+        // Exclude Classes.
+        case '{':
+          if ($inClass) {
+            $bracesCount++;
+          }
+          break;
+
+        case '}':
+          if ($inClass) {
+            $bracesCount--;
+            if ($bracesCount === 0) {
+              $inClass = FALSE;
+            }
+          }
+          break;
+      }
+    }
+
+    return $functions;
   }
 
 }
